--- logs/application.log
@@ -1,2 +1,0 @@
-2023-01-19 09:00:43.301  WARN 14920 --- [HikariPool-1 housekeeper] com.zaxxer.hikari.pool.HikariPool        : HikariPool-1 - Thread starvation or clock leap detected (housekeeper delta=15h14m3s276ms364µs200ns).
-2023-01-19 09:00:47.434  WARN 14920 --- [HikariPool-2 housekeeper] com.zaxxer.hikari.pool.HikariPool        : HikariPool-2 - Thread starvation or clock leap detected (housekeeper delta=15h14m3s309ms328µs).
